linstor-satellite/requirements.txt
ops >= 1.2.0
oci-image >= 1.0.0
kubernetes >= 17.17.0
python-linstor >= 1.8.0
